refactor(Tags): Reorganized files for Tag color
