+++ src/modules/tags/TagDetailSidePanel.js
@@ -1,18 +1,16 @@
 import React, {useState} from 'react';
-import {Text, View} from 'react-native';
+import {View} from 'react-native';
 
-import {Icon} from '@rn-vui/base';
 import {useDispatch, useSelector} from 'react-redux';
 
-import ColorPickerModal from './ColorPickerModal';
 import {isEmpty} from '../../shared/Helpers';
-import {SMALL_TEXT_SIZE} from '../../shared/styles.constants';
 import {MAIN_MENU_ITEMS, SIDE_PANEL_VIEWS} from '../main-menu-panel/mainMenu.constants';
 import {setSidePanelVisible} from '../main-menu-panel/mainMenuPanel.slice';
 import SidePanelHeader from '../main-menu-panel/sidePanel/SidePanelHeader';
 import {PAGE_KEYS} from '../page/pageKeys.constants';
 import {setSelectedAttributes, setSelectedSpot} from '../spots/spots.slice';
 import {TagDetail, TagDetailModal} from '../tags';
+import TagColor from './color/TagColor';
 
 const TagDetailSidePanel = ({openNotebookPanel}) => {
   /* Data Hooks */
@@ -22,7 +20,6 @@ const TagDetailSidePanel = ({openNotebookPanel}) => {
 
   /* Local State */
 
-  const [isColorPickerModalVisible, setIsColorPickerModalVisible] = useState(false);
   const [isDetailModalVisible, setIsDetailModalVisible] = useState(false);
 
   /* Derived Variables */
@@ -53,17 +50,7 @@ const TagDetailSidePanel = ({openNotebookPanel}) => {
             title={label}
           />
         </View>
-        <View style={{width: 100, position: 'absolute', right: 0, top: 0, alignItems: 'center'}}>
-          <Text style={{paddingBottom: 5, paddingTop: 5, fontSize: SMALL_TEXT_SIZE}}>{colorLabel} Color</Text>
-          <Icon
-            color={selectedTag.color}
-            containerStyle={{borderWidth: 1}}
-            name={selectedTag.color ? 'square' : 'x-square'}
-            onPress={() => setIsColorPickerModalVisible(true)}
-            size={30}
-            type={selectedTag.color ? 'ionicon' : 'feather'}
-          />
-        </View>
+        <TagColor colorLabel={colorLabel}/>
       </View>
 
       <View style={{flex: 1}}>
@@ -82,8 +69,9 @@ const TagDetailSidePanel = ({openNotebookPanel}) => {
           setIsDetailModalVisible={() => setIsDetailModalVisible(true)}
         />
       </View>
+
+      {/* Modals */}
       {isDetailModalVisible && <TagDetailModal closeModal={closeDetailModal}/>}
-      {isColorPickerModalVisible && <ColorPickerModal closeModal={() => setIsColorPickerModalVisible(false)}/>}
     </View>
   );
 };
