+++ src/modules/tags/TagsList.js
@@ -5,7 +5,7 @@ import {ListItem} from '@rn-vui/base';
 import {useDispatch, useSelector} from 'react-redux';
 
 import {useTags} from '.';
-import TagColorIcon from './TagColorIcon';
+import TagColorIcon from './color/TagColorIcon';
 import {TAG_SECTIONS, TAG_TYPES} from './tags.constants';
 import {getTagTitle} from './tags.helpers';
 import commonStyles from '../../shared/common.styles';
